Resolve Python linkage issues on mac (#247)
Re-work the python location code so we have more control on
macOS for how we find and link the python; automate the
manual step which moves python into the bundle for python-less
startup. Add several messages as to versions to both cmake
and opening bundle of messages.

Works on Lin and Win also but Win still has the DLL problem
reported by rovingeye.