More Mac Changes
Ahh the many names of python on dev boxes

Anyway this deals with a case where you pick up the python 3.8 which
ships inside xcode rather than the python you have installed, which
it seems happens on M1 macs because of non-standard homebrew
locations. This gives a warning and a hint of how to fix, but
also produces a runnable bespoke in that situation.