fix softlock due to windows messaging breaking down in certain scenarios (#310)
fixes #123
in windows, when certain events happened (plugging in a monitor, unplugging a monitor, etc), bespoke would stop receiving messages from windows, such as mouse clicks. this seems to have been caused by the juce audio device scanner spawning up a hidden window, but that window was not spawned up by the main thread, which seems to have caused issues. the fix is to ensure that that window is spawned by the main thread. thank you to @gamingrobot for the incredible work in investigating this.