Fixes performance issues on Linux
This should fix #309 which originated from #30, according to JUCE documentation if you're handling rendering yourself you should set `ComponentPaintingEnabled` to false which is the case with Bespoke, this might also benefit other platforms

```
By default this is set to true. If you're not using any paint() method functionality and are doing all your rendering in an OpenGLRenderer, you should disable it to improve performance.
```

restored debug directive