Fix and simplify resource directory discovery
The resource directory was resolved relative to the CWD everywhere
except on macOS, causing resources to not be found in many cases.

This removes reliance on the CWD and only looks in one well-defined path
(two on non-Apple UNIX). If the resource dir ain't there, that's a usage
error and we throw.

The paths are as follows:

macOS: <Bundle directory>/Contents/Resources/resource
Windows: <BespokeSynth.exe directory>/resource
UNIX: <BespokeSynth executable directory>/resource
      <prefix>/share/BespokeSynth/resource (for distros)