Disable build-time version info by default, fix build when disabled
The BESPOKE_RELIABLE_VERSION_INFO build option makes sure build info is
always up to date, but the flip side is that the build is always out of
date. This slows down debug sessions and can hide bugs in other CMake
rules that cause them to be always out of date unintentionally.

In a plugin (such as Surge, where this code came from), this tradeoff is
worth it because it's not always clear where the plugin gets loaded from
at any given time, and reliable build info by default can save lots of
grey hair.

But in a stand-alone application like Bespoke, it makes more sense to
default to a sane build environment that knows when stuff needs to be
rebuilt. If you're unsure what you're running, just look at the .EXE.