Change for Discussion nnot Merging
So I don't think this is mergable - putting it here as a sort of
draft PR. Using the profiler on a heavy granular patch I found
more time in ChannelBuffer::GetChannel still. So what I did here
(which we can propagate if you agree) is add a GetChannelUnchecked
whcih is const inline direct access to the float with no if and no
alloc, and then used that. It took about 5% out of the CPU time
in the audio thread.

I think there's other performance hits in the granulator path also
and lots of little things like clamps and maps which don't inline
and stuff but pushing this up to see what you think of the idea.

Since I haven't checked every client of Grain::Process I don't
know which ones explicitly or implicitly guanrantee their buffers
so I probably wouldn't merge this