Using more precise "center" value for incremental sliders
Using an Arturia Minilab MkII, turning an endless encoder to the left
slightly sends a MIDI CC value of 63, seen in Bespoke as 0.50, and
doesn't effect any change in the value it's mapped to control.
I imagine this is caused by .5f being a less precise "center" value
than is needed in this case. Using 64/127.0f instead, the issue is
fixed.
The incrementThreshold in the Minilab's layout also needs to be set to
0 so as to not block any value from having an effect.