Fix notecanvas connections not disconnecting
IDrawableModule::KeyPressed needs to be called for the backspace
on the cable to work.
The canvas will still get the keypresses this way.
The `GetLastClickedModule` condition was removed, because
drag-selecting notes in the canvas will not set that properly.
So just placing the mouse over a canvas will allow the key interaction.