update slider smoother UI
previously, all sliders would switch to "smooth" mode when holding alt, which was kinda ugly.
now, they only switch to that UI if you hold alt and hover over them, or if they are already smoothing. this also gives the alt key some nice functionality of "reveal to me which sliders are smoothed"