linux fps: undo #396 and restore #349 for linux only (#505)
after testing, I and others can't reproduce the thread safety crash of #396 on linux, only on mac and windows. something must be going on under the juce hood with threading, to make this setting work for linux with my thread model, and bad for mac/windows.

long-term the real issue is safety issues in my thread model, so that should be fixed. but this seems to be safe in the short-term.