Xoshiro256ss: rename .h + add std::random_device ctor
I accidentally used the wrong name :( My bad. Anyways:

The new constructor properly seeds the PRNG fully (to 256 bits) when used, which should be better. It might be a tiny bit slower (we go from one fetch of the random_device to 2*4 = 8), however we only do seeding on construction, So it should be fine. It's not like it's 5000 bytes of state either.

Changes SynthGlobals.cpp to use the newly introduced random_device& ctor.