add option for hockets to be repeatably random
notehocket and pulsehocket send notes and pulses to random destinations, and now they have a new `deterministic` mode where you can set a random seed and are able to reset the random index, to make them random but repeatable in loops. this repetition makes the randomization feel much more musical!