change default zoom level and ui scale of bespoke to 1.3 (#983)
I thought I had done this months ago! whoops. won't affect users who already have launched bespoke, but when a userprefs.json is created, it will use 1.3 as the zoom level, which is how I've been using bespoke for the past few months.

also snuck in a change to the default scenes for songbuilder