Add grid connections only after layout is loaded (#1173)
When a controller layout other than the one named exactly like the midi
device is used in a saved state, connections on pages that were not
active when saving are not restored.

This happens because, the grid connections were added in Init.
At that point, the controller uses the default layout (i.e. the one
named like the device or the actual default without any grids).
Afterwards the UI State is loaded in LoadState, and then the selection
from the Layout dropdown is restored causing mGrids to be (re-)set
to the correct setup.

Now, we load the grids in LoadState after the UI has restored the
state of the layout dropdown, so grid connections are always loaded
into the correct grids.