fix crash on load caused by effectchains with deleted prior effects of the same type (#1357)
when adding effects to an effect chain, numbers are automatically appended to avoid collisions (like, if you spawn a granulator in an effectchain, it will be called "granulator", and if you spawn another, it will be called "granulator2"). the effect names were not being loaded in properly, so if you deleted the first granulator, you'd just be left with "granulator2", but on load it would automatically name it "granulator", and all references to "granulator2" would be busted, causing a crash on load. fixes #1354

***NO_CI***