Added NoteSetVoice node, Debug VST assert fix
Added a new node that allows setting a specific voice channel on a note. Combined with Use Voice As Channel on the PST plugin node, this allows the use of a single Kontakt vst (for example) and send different notes to different instruments.

Fixed an issue in VSTPlayhead where an optional time signature node was causing and assert in debug builds