Move Bespoke to JUCE 7.0.12
There's a bug in JUCE 7.0.7 and earlier

https://github.com/juce-framework/JUCE/commit/335f6e9591f6f09f74e13769c74b2d345e8336c9

which means plugins which query for host side menus can get a memory
error. Surge XT is such a plugin. Therefore with JUCE 7.0.7 Surge XT
will crash in bespoke.

Moving beyond that commit (which was between 7.0.7 and 7.0.8) fixes it
but if you are moving might as well move to the last of the 7 series with
7.0.12, which this PR does