Add FluidSynth module
This implements a native FluidSynth based module. A SoundFont can be
specified and the presets/instruments for each channel can be selected.

The FluidSynth code is added as another submodule under libs but the
build can be done in either of three ways:

* BESPOKE_FLUIDSYNTH=OFF to turn off the FluidSynth module entirely
* BESPOKE_SYSTEM_FLUIDSYNTH=ON to use a system provided library
* BESPOKE_SYSTEM_FLUIDSYNTH=OFF (default) to build from internal source

Using the system provided FluidSynth has the benefit of it possibly
being configured to find a default SoundFont automatically so the module
would work out of the box. Building from source on the other hand allows
for a known version to be used (2.4.5 in this commit). When building
from source, a stripped down library is built that excludes any audio
drivers or MIDI IO as that is not needed here.