FluidSynth: Fix missing controller change messages.
First, SendCC is not called for these by default but requires the
MIDIController to be configured with sendoutputcc. So still forward
them in SendMidi.

Second, message.isController() only checks if the message is a
controller message (0xbX) and therefore also covers the more specific
status codes. It therefore has to be checked last.