Add NamedMutex::TryLock() and ScopedTryMutex using it.
They allow to not block in case a lock is currently held. It is meant to
be used on the audio thread when something needs to be locked to work
but must not block to wait for the lock to be released.

The mode of operation is to use NamedMutex::TryLock() and check its
return value or use the ScopedTryMutex and check its IsLocked() return
value to know whether or not the lock was acquired. If acquisition
failed, the shared resources protected by the mutex cannot be used and
some fallback should take place.