Make layout less tall by using two columns for preset dropdowns.
It's usually the vertical space that is constrained on screen, so a less
tall layout makes it easier to work with the module. Also drop the
automatic size of the path TextEntry so the module does not become
arbitrarily wide. The path is less critical now that there is a file
open dialog.