Add Behringer BCR2000 controller definitions
BCR2000.json

	Matches the layout of the front pannel controls that emit MIDI
	plus two foot switch inputs, factory default first preset, first
	encoder group only.

BCR2000_encoder_groups.json

	The same except for having four copies of the top row of
	encoders, selectable with the four encoder group buttons on the
	hardware.

If you are not using multiple encoder groups then the "BCR2000.json"
layout is a bit more compact and visually analogous to the hardware.

In both layouts the button directly above each knob in the first row of
knobs represents the push function of that encoder. The other three rows
of encoders have no such function in the hardware.